fix sign-conversion warnings for metal samples

+++ metal/osxentry.m
@@ -109,7 +109,7 @@ static id mtk_view_controller;
     [mtk_view setDevice: mtl_device];
     [mtk_view setColorPixelFormat:MTLPixelFormatBGRA8Unorm];
     [mtk_view setDepthStencilPixelFormat:MTLPixelFormatDepth32Float_Stencil8];
-    [mtk_view setSampleCount:sample_count];
+    [mtk_view setSampleCount:(NSUInteger)sample_count];
     #if !TARGET_OS_IPHONE
         [window setContentView:mtk_view];
         CGSize drawable_size = { (CGFloat) width, (CGFloat) height };
